[📦] bumped kord-ex verion + calc commands added

+++ src/main/kotlin/arguments/CalcTElem.kt
@@ -1,10 +1,11 @@
 package arguments
 
+import com.kotlindiscord.kord.extensions.commands.converters.impl.decimal
 import com.kotlindiscord.kord.extensions.commands.converters.impl.int
 import com.kotlindiscord.kord.extensions.commands.parser.Arguments
 import dev.kord.common.annotation.KordPreview
 
 @KordPreview
 class CalcTElem: Arguments() {
-    val attack by int("attack", "in-game attack value")
+    val attack by decimal("attack", "in-game attack value")
 }
